split web into individual pages
